Switch over to torch_dtype

+++ train_sd1.py
@@ -323,8 +323,7 @@ def main():
     accelerator.print("Now loading core model files.")
     noise_scheduler = DDPMScheduler.from_pretrained(settings["model_name"], subfolder="scheduler")
     unet = UNet2DConditionModel.from_pretrained(settings["model_name"], subfolder="unet", main_dtype=main_dtype)
-    text_model = CLIPTextModel.from_pretrained(settings["model_name"], subfolder="text_encoder")
-    text_model = text_model.to(dtype=main_dtype)
+    text_model = CLIPTextModel.from_pretrained(settings["model_name"], subfolder="text_encoder", torch_dtype=main_dtype)
     if settings["enable_gradient_checkpointing"]:
         unet.enable_gradient_checkpointing()
         if settings["train_text_encoder"]:
